refactor(todo): P7 统一 Reorderable 拖拽模式为便捷方法
任务：
把 [CheckboxEditText.kt] 中 2 个 Modifier.draggableHandle(..., LongPress)
改造为 Modifier.longPressDraggableHandle(...)，与项目其他 5 个拖拽点
（ZonedReorderableLazyColumn + ReorderableLazyColumn + SpecialDateScreen）
的形式保持一致。

改造动机：
- 与旧代码 `Modifier.draggableHandle(..., dragGestureDetector = LongPress)`
  完全等价，源码实现就是便捷封装
- 项目统一用便捷方法后，意图明确（永远要长按），无需担心默认值
- 从源头杜绝"忘记传 dragGestureDetector 导致 Press 误触"的低级错误

具体改动：
1. [CheckboxEditText.kt:1425] 图片部分：Modifier.draggableHandle → longPressDraggableHandle
2. [CheckboxEditText.kt:1550] 语音部分：Modifier.draggableHandle → longPressDraggableHandle
3. [CheckboxEditText.kt:70] 删除 import sh.calvin.reorderable.DragGestureDetector
4. 改写两处 KDoc 注释（P5/P6 改造动机 → P7 便捷方法说明）

P7 改造后项目状态：
- 7/7 拖拽点全部使用 longPressDraggableHandle 便捷方法
- Press 模式残留：0 处
- DragGestureDetector 引用：0 处（仅 KDoc 历史注释中提到）
- 功能行为：完全无变化（仅形式统一）

后续可能：
- 编译验证建议：下次会话开始跑 ./gradlew assembleDebug
- Reorderable 库使用规范第 8 节建议更新使用统计（5/7 → 7/7）
- 后续可删除 P5/P6 历史注释中的 dragGestureDetector 引用

+++ app/src/main/java/com/corgimemo/app/ui/components/calendar/DatePickerRow.kt
@@ -130,10 +130,10 @@ fun DatePickerRow(
         val dayW = dayP.width
         val arrowW = arrowP.width
 
-        // 间距（dp → px）
-        val gapMonthInnerPx = 0.dp.toPx()    // "07" → "月" 0dp（视觉上"07月"连续）
-        val gapMonthDayPx = 7.dp.toPx()      // "月" → "27" 7dp
-        val gapDayArrowPx = 2.dp.toPx()      // "27" → "▼" 2dp
+        // 间距（dp → px，toInt() 避免 Int + Float = Float 的类型错误）
+        val gapMonthInnerPx = 0.dp.toPx().toInt()    // "07" → "月" 0dp（视觉连续）
+        val gapMonthDayPx = 7.dp.toPx().toInt()      // "月" → "27" 7dp
+        val gapDayArrowPx = 2.dp.toPx().toInt()      // "27" → "▼" 2dp
 
         val totalW = monthNumW + gapMonthInnerPx + monthCnW + gapMonthDayPx + dayW + gapDayArrowPx + arrowW
